Make a mixin to accomodate decimal
