+++ tsepamo/models/model_mixins/__init__.py
@@ -3,3 +3,4 @@ from .complete_field_mixin import CompleteFieldMixin
 from .outcomes_model_mixin import OutcomesModelMixin
 from .uuid_model_mixin import UuidModelMixin
 from .project_field_mixin import ProjectFieldMixin
+from .decimal_field_mixin import DecimalFieldMixin
